requirements.txt
fastapi==0.95.2
httpx==0.24.1
uvicorn==0.22.0
streamlit==1.30.0
pandas==2.2.2
numpy==1.26.2
scikit-learn==1.3.2
statsmodels==0.14.1
requests==2.31.0
pyyaml==6.0
pytest==7.4.2
geopy==2.4.0
folium==0.14.0
xarray==2024.11.0
python-dotenv==1.0.0
python-multipart==0.0.9
joblib==1.3.2
streamlit-folium==0.18.0
